# src/Controllers/LoggerController.cs
/*
 * IConfiguration and IOption test
 *
 * Test API to show using IOptions and IConfiguration
 *
 * OpenAPI spec version: 1.0.0-oas3
 * Contact: myemail@aisreview.com
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;

using Microsoft.AspNetCore.Authorization;
using IOptionTest;
using System.Threading.Tasks;

namespace IOptionTest.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class LoggerApiController : ControllerBase
    { 
        /// <summary>
        /// Log a message
        /// </summary>
        /// <param name="body">Message to log</param>
        /// <response code="200">Ok</response>
        /// <response code="400">bad input parameter</response>
        [HttpPost]
        [Route("/api/logger")]
        [ValidateModelState]
        [SwaggerOperation("LogMessage")]
        public virtual IActionResult LogMessage([FromBody]Message body)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);

            throw new NotImplementedException();
        }
    }
}